implement the check alive on the sim

+++ src/main/libautodiag/buffer.c
@@ -386,15 +386,15 @@ uint8_t ad_buffer_to_be8(Buffer * buffer) {
 }
 uint16_t ad_buffer_to_be16(Buffer * buffer) {
     assert(2 <= buffer->size);
-    return (buffer->buffer[0] << 8) | buffer->buffer[1];
+    return ((uint16_t)buffer->buffer[0] << 8) | buffer->buffer[1];
 }
 uint32_t ad_buffer_to_be32(Buffer * buffer) {
     assert(4 <= buffer->size);
-    return (buffer->buffer[0] << (8 * 3)) | (buffer->buffer[1] << (8 * 2)) | (buffer->buffer[2] << 8) | buffer->buffer[3];
+    return ((uint32_t)buffer->buffer[0] << (8 * 3)) | ((uint32_t)buffer->buffer[1] << (8 * 2)) | ((uint32_t)buffer->buffer[2] << 8) | buffer->buffer[3];
 }
 uint64_t ad_buffer_to_be64(Buffer * buffer) {
     assert(8 <= buffer->size);
-    return (buffer->buffer[0] << (8 * 7)) | (buffer->buffer[1] << (8 * 6)) | (buffer->buffer[2] << (8 * 5)) | (buffer->buffer[3] << (8 * 4)) | (buffer->buffer[4] << (8 * 3)) | (buffer->buffer[5] << (8 * 2)) | (buffer->buffer[6] << 8) | buffer->buffer[7];
+    return ((uint64_t)buffer->buffer[0] << (8 * 7)) | ((uint64_t)buffer->buffer[1] << (8 * 6)) | ((uint64_t)buffer->buffer[2] << (8 * 5)) | ((uint64_t)buffer->buffer[3] << (8 * 4)) | ((uint64_t)buffer->buffer[4] << (8 * 3)) | ((uint64_t)buffer->buffer[5] << (8 * 2)) | ((uint64_t)buffer->buffer[6] << 8) | buffer->buffer[7];
 }
 char * ad_buffer_to_ascii(final Buffer *buffer) {
     if ( buffer == null ) {
